+++ deploys/postgres_upgrade.py
@@ -1,0 +1,5 @@
+import os
+
+from server_setup import postgres_upgrade
+
+postgres_upgrade(confirm=os.environ.get("CONFIRM") == "1", drop_old=os.environ.get("DROP_OLD") == "1")